LPS-205289 Rename Objects playwright test folder from object to object-web
