+++ modules/test/playwright/playwright.config.ts
@@ -6,7 +6,7 @@
 import {defineConfig} from '@playwright/test';
 
 import {config as setup} from './tests/global.setup.config';
-import {config as object} from './tests/object/config';
+import {config as object} from './tests/object-web/config';
 import {config as portalWeb} from './tests/portal-web/config';
 
 export default defineConfig({
