+++ modules/test/playwright/tests/object-web/config.ts
@@ -8,7 +8,7 @@ import {devices} from '@playwright/test';
 export const config = {
 	dependencies: ['setup'],
 	name: 'object',
-	testDir: 'tests/object',
+	testDir: 'tests/object-web',
 	use: {
 		...devices['Desktop Chrome'],
 		storageState: 'tmp/.auth/user.json',
